renamed claims controller to cases controller
