+++ src/main/java/uk/gov/hmcts/reform/civil/controllers/cases/CasesController.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.controllers.claims;
+package uk.gov.hmcts.reform.civil.controllers.cases;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -23,21 +23,21 @@ import uk.gov.hmcts.reform.civil.service.CoreCaseDataService;
     path = "/cases",
     produces = MediaType.APPLICATION_JSON_VALUE
 )
-public class ClaimsController {
+public class CasesController {
 
     private final CoreCaseDataService coreCaseDataService;
     private final CaseDetailsConverter caseDetailsConverter;
 
     @GetMapping(path = {
-        "/{claimId}",
+        "/{cid}",
     })
-    @ApiOperation("get claim by id from CCD")
+    @ApiOperation("get case by id from CCD")
 
     public ResponseEntity<CaseData> getClaimById(
-        @PathVariable("claimId") Long claimId
+        @PathVariable("cid") Long claimId
     ) {
         log.info(
-            "Received ClaimId: {}",
+            "Received CaseId: {}",
             claimId
         );
